sorted_pset_trials - fixed typing in particle file writer for vectorized version

+++ parcels/particlefile_vectorized.py
@@ -341,11 +341,11 @@ class ParticleFile(object):
         """
 
         data = np.nan * np.zeros((self.maxid_written+1, time_steps))
-        time_index = np.zeros(self.maxid_written+1, dtype=int)
+        time_index = np.zeros(self.maxid_written+1, dtype=np.int32)
 
         # data = np.nan * np.zeros((self.max_index_written+1, time_steps))
         # time_index = np.zeros(self.max_index_written+1, dtype=int)
-        t_ind_used = np.zeros(time_steps, dtype=int)
+        t_ind_used = np.zeros(time_steps, dtype=np.int32)
 
         # loop over all files
         for npyfile in file_list:
